refactor(perl-lsp-rs): split code-actions provider into SRP submodules (#0000)
